requirement.txt
Django==2.1.5
django-cors-headers==3.0.1
django-debugging==0.1a2
django-filter==2.1.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
mysqlclient==1.4.2.post1
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
